+++ app/src/main/java/com/jerboa/ui/components/post/PostListing.kt
@@ -1,39 +1,22 @@
 package com.jerboa.ui.components.post
 
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.combinedClickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.text.selection.SelectionContainer
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.outlined.Comment
-import androidx.compose.material.icons.filled.Bookmark
-import androidx.compose.material.icons.outlined.BookmarkBorder
-import androidx.compose.material.icons.outlined.ChatBubbleOutline
-import androidx.compose.material.icons.outlined.CommentsDisabled
-import androidx.compose.material.icons.outlined.Delete
-import androidx.compose.material.icons.outlined.Gavel
 import androidx.compose.material.icons.outlined.Link
-import androidx.compose.material.icons.outlined.MoreVert
-import androidx.compose.material.icons.outlined.PushPin
 import androidx.compose.material3.Card
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
-import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedCard
 import androidx.compose.material3.Text
@@ -41,1044 +24,50 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontStyle
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
 import com.jerboa.JerboaAppState
 import com.jerboa.PostType
 import com.jerboa.PostViewMode
 import com.jerboa.R
 import com.jerboa.datatypes.BanFromCommunityData
 import com.jerboa.datatypes.PostFeatureData
-import com.jerboa.datatypes.sampleImagePostView
-import com.jerboa.datatypes.sampleLinkNoThumbnailPostView
-import com.jerboa.datatypes.sampleLinkPostView
-import com.jerboa.datatypes.sampleMarkdownPostView
-import com.jerboa.datatypes.samplePostView
 import com.jerboa.db.entity.Account
-import com.jerboa.db.entity.AnonAccount
 import com.jerboa.feat.BlurNSFW
 import com.jerboa.feat.InstantScores
 import com.jerboa.feat.PostActionBarMode
 import com.jerboa.feat.SwipeToActionPreset
 import com.jerboa.feat.SwipeToActionType
 import com.jerboa.feat.VoteType
-import com.jerboa.feat.amMod
-import com.jerboa.feat.canMod
-import com.jerboa.feat.default
 import com.jerboa.feat.isReadyAndIfNotShowSimplifiedInfoToast
-import com.jerboa.feat.needBlur
-import com.jerboa.feat.simulateModerators
 import com.jerboa.getPostType
-import com.jerboa.hostName
-import com.jerboa.isSameInstance
-import com.jerboa.nsfwCheck
-import com.jerboa.rememberJerboaAppState
-import com.jerboa.siFormat
-import com.jerboa.toHttps
-import com.jerboa.ui.components.common.ActionBarButton
-import com.jerboa.ui.components.common.ActionBarButtonAndBadge
-import com.jerboa.ui.components.common.CircularIcon
-import com.jerboa.ui.components.common.CommentOrPostNodeHeader
-import com.jerboa.ui.components.common.DotSpacer
-import com.jerboa.ui.components.common.MarkdownHelper.CreateMarkdownPreview
 import com.jerboa.ui.components.common.MyMarkdownText
-import com.jerboa.ui.components.common.NsfwBadge
 import com.jerboa.ui.components.common.PictrsThumbnailImage
-import com.jerboa.ui.components.common.PictrsUrlImage
-import com.jerboa.ui.components.common.ScoreAndTime
-import com.jerboa.ui.components.common.SimpleTopAppBar
 import com.jerboa.ui.components.common.SwipeToAction
-import com.jerboa.ui.components.common.TimeAgo
-import com.jerboa.ui.components.common.VoteGeneric
-import com.jerboa.ui.components.common.fadingEdge
 import com.jerboa.ui.components.common.rememberSwipeActionState
-import com.jerboa.ui.components.common.scoreColor
-import com.jerboa.ui.components.community.CommunityLink
-import com.jerboa.ui.components.community.CommunityName
-import com.jerboa.ui.components.person.PersonProfileLink
-import com.jerboa.ui.components.post.composables.PostOptionsDropdown
 import com.jerboa.ui.components.settings.about.TORRENT_HELP_LINK
-import com.jerboa.ui.theme.ACTION_BAR_ICON_SIZE
-import com.jerboa.ui.theme.LARGER_ICON_THUMBNAIL_SIZE
 import com.jerboa.ui.theme.LARGE_PADDING
 import com.jerboa.ui.theme.LINK_ICON_SIZE
-import com.jerboa.ui.theme.MEDIUM_ICON_SIZE
 import com.jerboa.ui.theme.MEDIUM_PADDING
 import com.jerboa.ui.theme.POST_LINK_PIC_SIZE
-import com.jerboa.ui.theme.SMALLER_PADDING
-import com.jerboa.ui.theme.SMALL_PADDING
 import com.jerboa.ui.theme.THUMBNAIL_CARET_SIZE
-import com.jerboa.ui.theme.XXL_PADDING
 import com.jerboa.ui.theme.jerboaColorScheme
-import it.vercruysse.lemmyapi.datatypes.*
-import kotlinx.coroutines.CoroutineScope
+import it.vercruysse.lemmyapi.datatypes.Community
+import it.vercruysse.lemmyapi.datatypes.LocalUserVoteDisplayMode
+import it.vercruysse.lemmyapi.datatypes.Person
+import it.vercruysse.lemmyapi.datatypes.PersonId
+import it.vercruysse.lemmyapi.datatypes.PersonView
+import it.vercruysse.lemmyapi.datatypes.Post
+import it.vercruysse.lemmyapi.datatypes.PostId
+import it.vercruysse.lemmyapi.datatypes.PostView
 
-@Composable
-fun PostHeaderLine(
-    community: Community,
-    post: Post,
-    creator: Person,
-    creatorBannedFromCommunity: Boolean,
-    instantScores: InstantScores,
-    onCommunityClick: (community: Community) -> Unit,
-    onPersonClick: (personId: PersonId) -> Unit,
-    modifier: Modifier = Modifier,
-    showCommunityName: Boolean = true,
-    showAvatar: Boolean,
-    fullBody: Boolean,
-    blurNSFW: BlurNSFW,
-    voteDisplayMode: LocalUserVoteDisplayMode,
-) {
-    Column(modifier = modifier) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(MEDIUM_PADDING),
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier.weight(1f),
-            ) {
-                if (showCommunityName && showAvatar) {
-                    CommunityIcon(community, onCommunityClick, blurNSFW)
-                }
-                Column {
-                    if (showCommunityName) {
-                        CommunityName(
-                            community = community,
-                            onClick = { onCommunityClick(community) },
-                        )
-                    }
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(SMALL_PADDING),
-                    ) {
-                        PersonProfileLink(
-                            person = creator,
-                            onClick = onPersonClick,
-                            showTags = fullBody,
-                            // Set this to false, we already know this
-                            isPostCreator = false,
-                            isCommunityBanned = creatorBannedFromCommunity,
-                            color = MaterialTheme.colorScheme.outline,
-                            showAvatar = !showCommunityName && showAvatar,
-                        )
-                        if (post.featured_local) {
-                            DotSpacer()
-                            Icon(
-                                imageVector = Icons.Outlined.PushPin,
-                                contentDescription = stringResource(R.string.postListing_featuredLocal),
-                                tint = MaterialTheme.colorScheme.primary,
-                                modifier = Modifier.size(ACTION_BAR_ICON_SIZE),
-                            )
-                        }
-                        if (post.featured_community) {
-                            DotSpacer()
-                            Icon(
-                                imageVector = Icons.Outlined.PushPin,
-                                contentDescription = stringResource(R.string.postListing_featuredCommunity),
-                                tint = MaterialTheme.colorScheme.secondary,
-                                modifier = Modifier.size(ACTION_BAR_ICON_SIZE),
-                            )
-                        }
-                        if (post.locked) {
-                            DotSpacer()
-                            Icon(
-                                imageVector = Icons.Outlined.CommentsDisabled,
-                                contentDescription = stringResource(R.string.postListing_locked),
-                                tint = MaterialTheme.colorScheme.error,
-                                modifier = Modifier.size(ACTION_BAR_ICON_SIZE),
-                            )
-                        }
-                    }
-                }
-                if (post.deleted) {
-                    Icon(
-                        imageVector = Icons.Outlined.Delete,
-                        contentDescription = stringResource(R.string.postListing_deleted),
-                        tint = MaterialTheme.colorScheme.error,
-                    )
-                }
-                if (post.removed) {
-                    Icon(
-                        imageVector = Icons.Outlined.Gavel,
-                        contentDescription = stringResource(R.string.removed),
-                        tint = MaterialTheme.colorScheme.error,
-                    )
-                }
-            }
-            ScoreAndTime(
-                instantScores = instantScores,
-                published = post.published,
-                updated = post.updated,
-                isNsfw = nsfwCheck(post, community),
-                voteDisplayMode = voteDisplayMode,
-            )
-        }
-    }
-}
-
-@Preview
-@Composable
-fun PostHeaderLinePreview() {
-    val postView = sampleLinkPostView
-    PostHeaderLine(
-        post = postView.post,
-        community = postView.community,
-        creator = postView.creator,
-        creatorBannedFromCommunity = postView.creator_banned_from_community,
-        instantScores = InstantScores(
-            myVote = 0,
-            score = 10,
-            upvotes = 9,
-            downvotes = 1,
-        ),
-        onCommunityClick = {},
-        onPersonClick = {},
-        showAvatar = true,
-        blurNSFW = BlurNSFW.NSFW,
-        fullBody = true,
-        voteDisplayMode = LocalUserVoteDisplayMode.default(),
-    )
-}
-
-@Composable
-fun CommunityIcon(
-    community: Community,
-    onCommunityClick: (community: Community) -> Unit,
-    blurNSFW: BlurNSFW,
-) {
-    community.icon?.let {
-        CircularIcon(
-            icon = it,
-            contentDescription = stringResource(R.string.postListing_goToCommunity),
-            size = MEDIUM_ICON_SIZE,
-            modifier = Modifier.clickable { onCommunityClick(community) },
-            thumbnailSize = LARGER_ICON_THUMBNAIL_SIZE,
-            blur = blurNSFW.needBlur(community.nsfw),
-        )
-    }
-}
-
-@Composable
-fun PostNodeHeader(
-    postView: PostView,
-    instantScores: InstantScores,
-    onPersonClick: (personId: PersonId) -> Unit,
-    showAvatar: Boolean,
-    voteDisplayMode: LocalUserVoteDisplayMode,
-) {
-    CommentOrPostNodeHeader(
-        creator = postView.creator,
-        instantScores = instantScores,
-        published = postView.post.published,
-        updated = postView.post.updated,
-        deleted = postView.post.deleted,
-        onPersonClick = onPersonClick,
-        isPostCreator = true,
-        isCommunityBanned = postView.creator_banned_from_community,
-        onClick = {},
-        onLongCLick = {},
-        showAvatar = showAvatar,
-        voteDisplayMode = voteDisplayMode,
-        isDistinguished = false,
-    )
-}
-
-@Composable
-fun PostTitleBlock(
-    post: Post,
-    read: Boolean,
-    expandedImage: Boolean,
-    account: Account,
-    useCustomTabs: Boolean,
-    usePrivateTabs: Boolean,
-    blurEnabled: Boolean,
-    appState: JerboaAppState,
-    showIfRead: Boolean,
-) {
-    val imagePost = post.url?.let { getPostType(it) == PostType.Image } ?: false
-
-    if (imagePost && expandedImage) {
-        PostTitleAndImageLink(
-            post = post,
-            read = read,
-            appState = appState,
-            blurEnabled = blurEnabled,
-            showIfRead = showIfRead,
-        )
-    } else {
-        PostTitleAndThumbnail(
-            post = post,
-            read = read,
-            account = account,
-            useCustomTabs = useCustomTabs,
-            usePrivateTabs = usePrivateTabs,
-            blurEnabled = blurEnabled,
-            appState = appState,
-            showIfRead = showIfRead,
-        )
-    }
-}
-
-@Composable
-fun PostName(
-    post: Post,
-    read: Boolean,
-    showIfRead: Boolean,
-) {
-    val color =
-        if (showIfRead && read) {
-            MaterialTheme.colorScheme.outline
-        } else if (post.featured_local) {
-            MaterialTheme.colorScheme.primary
-        } else if (post.featured_community) {
-            MaterialTheme.colorScheme.secondary
-        } else {
-            Color.Unspecified
-        }
-
-    Text(
-        text = post.name,
-        style = MaterialTheme.typography.headlineMedium,
-        color = color,
-        modifier = Modifier.testTag("jerboa:posttitle"),
-    )
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun PostTitleAndImageLink(
-    post: Post,
-    read: Boolean,
-    blurEnabled: Boolean,
-    appState: JerboaAppState,
-    showIfRead: Boolean,
-) {
-    // This was tested, we know it exists
-    val url = post.url?.toHttps()
-
-    Column(
-        modifier = Modifier.padding(
-            vertical = MEDIUM_PADDING,
-            horizontal = MEDIUM_PADDING,
-        ),
-    ) {
-        // Title of the post
-        PostName(
-            post = post,
-            read = read,
-            showIfRead = showIfRead,
-        )
-    }
-
-    url?.let { cUrl ->
-        PictrsUrlImage(
-            url = cUrl,
-            blur = blurEnabled,
-            contentDescription = post.alt_text,
-            modifier =
-                Modifier
-                    .combinedClickable(
-                        onClick = { appState.openImageViewer(cUrl) },
-                        onLongClick = { appState.showLinkPopup(cUrl) },
-                    ),
-        )
-    }
-}
-
-@Composable
-fun PostTitleAndThumbnail(
-    post: Post,
-    read: Boolean,
-    account: Account,
-    useCustomTabs: Boolean,
-    usePrivateTabs: Boolean,
-    blurEnabled: Boolean,
-    appState: JerboaAppState,
-    showIfRead: Boolean,
-) {
-    Column(
-        modifier = Modifier.padding(horizontal = MEDIUM_PADDING),
-    ) {
-        Row(
-            horizontalArrangement = Arrangement.spacedBy(SMALL_PADDING),
-        ) {
-            // Title of the post
-            Column(
-                verticalArrangement = Arrangement.spacedBy(MEDIUM_PADDING),
-                modifier = Modifier.weight(1f),
-            ) {
-                PostName(post = post, read = read, showIfRead = showIfRead)
-                post.url?.also { postUrl ->
-                    if (!isSameInstance(postUrl, account.instance)) {
-                        val hostName = hostName(postUrl)
-                        hostName?.also {
-                            Text(
-                                text = it,
-                                color = MaterialTheme.colorScheme.outline,
-                                style = MaterialTheme.typography.labelMedium,
-                                fontFamily = FontFamily.Monospace,
-                            )
-                        }
-                    }
-                }
-            }
-            ThumbnailTile(
-                post = post,
-                useCustomTabs = useCustomTabs,
-                usePrivateTabs = usePrivateTabs,
-                blurEnabled = blurEnabled,
-                appState = appState,
-            )
-        }
-    }
-}
-
-@Composable
-fun PostBody(
-    post: Post,
-    read: Boolean,
-    fullBody: Boolean,
-    viewSource: Boolean,
-    expandedImage: Boolean,
-    account: Account,
-    useCustomTabs: Boolean,
-    usePrivateTabs: Boolean,
-    blurEnabled: Boolean,
-    showPostLinkPreview: Boolean,
-    appState: JerboaAppState,
-    clickBody: () -> Unit = {},
-    showIfRead: Boolean,
-) {
-    Column(
-        verticalArrangement = Arrangement.spacedBy(MEDIUM_PADDING),
-    ) {
-        PostTitleBlock(
-            post = post,
-            read = read,
-            expandedImage = expandedImage,
-            account = account,
-            useCustomTabs = useCustomTabs,
-            usePrivateTabs = usePrivateTabs,
-            blurEnabled = blurEnabled,
-            appState = appState,
-            showIfRead = showIfRead,
-        )
-
-        // The metadata card
-        if (fullBody && showPostLinkPreview) {
-            MetadataCard(post = post)
-
-            if (post.url?.startsWith("magnet") == true) {
-                TorrentHelpInfo()
-            }
-        }
-
-        // Check to make sure body isn't empty string
-        val body = post.body?.trim()?.ifEmpty { null }
-
-        // The desc
-        body?.also { text ->
-            if (fullBody) {
-                Column(
-                    modifier =
-                        Modifier
-                            .padding(MEDIUM_PADDING),
-                ) {
-                    if (viewSource) {
-                        SelectionContainer {
-                            Text(
-                                text = text,
-                                fontFamily = FontFamily.Monospace,
-                            )
-                        }
-                    } else {
-                        MyMarkdownText(
-                            markdown = text,
-                            onClick = {},
-                        )
-                    }
-                }
-            } else {
-                val bottomFade = Brush.verticalGradient(.7f to MaterialTheme.colorScheme.background, 1f to Color.Transparent)
-                CreateMarkdownPreview(
-                    markdown = text,
-                    color = LocalContentColor.current,
-                    onClick = clickBody,
-                    style = MaterialTheme.typography.bodyMedium,
-                    modifier = Modifier
-                        .padding(MEDIUM_PADDING)
-                        .fadingEdge(bottomFade),
-                )
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun PreviewStoryTitleAndMetadata() {
-    PostBody(
-        post = samplePostView.post,
-        read = samplePostView.read,
-        fullBody = false,
-        viewSource = false,
-        expandedImage = false,
-        account = AnonAccount,
-        useCustomTabs = false,
-        usePrivateTabs = false,
-        blurEnabled = BlurNSFW.NSFW.needBlur(samplePostView),
-        showPostLinkPreview = true,
-        appState = rememberJerboaAppState(),
-        showIfRead = true,
-    )
-}
-
-@Preview
-@Composable
-fun PreviewSourcePost() {
-    val pv = sampleMarkdownPostView
-    PostBody(
-        post = pv.post,
-        read = pv.read,
-        fullBody = true,
-        viewSource = true,
-        expandedImage = false,
-        account = AnonAccount,
-        useCustomTabs = false,
-        usePrivateTabs = false,
-        blurEnabled = BlurNSFW.NSFW.needBlur(pv),
-        showPostLinkPreview = true,
-        appState = rememberJerboaAppState(),
-        showIfRead = true,
-    )
-}
-
-@Composable
-fun PostFooterLine(
-    postView: PostView,
-    admins: List<PersonView>,
-    moderators: List<PersonId>?,
-    instantScores: InstantScores,
-    onUpvoteClick: () -> Unit,
-    onDownvoteClick: () -> Unit,
-    onReplyClick: (postView: PostView) -> Unit,
-    onSaveClick: (postView: PostView) -> Unit,
-    onEditPostClick: (postView: PostView) -> Unit,
-    onDeletePostClick: (postView: PostView) -> Unit,
-    onHidePostClick: (postView: PostView) -> Unit,
-    onReportClick: (postView: PostView) -> Unit,
-    onRemoveClick: (postView: PostView) -> Unit,
-    onBanPersonClick: (person: Person) -> Unit,
-    onBanFromCommunityClick: (banData: BanFromCommunityData) -> Unit,
-    onLockPostClick: (postView: PostView) -> Unit,
-    onFeaturePostClick: (data: PostFeatureData) -> Unit,
-    onViewVotesClick: (PostId) -> Unit,
-    onCommunityClick: (community: Community) -> Unit,
-    onPersonClick: (personId: PersonId) -> Unit,
-    onViewSourceClick: () -> Unit,
-    modifier: Modifier = Modifier,
-    showReply: Boolean = false,
-    account: Account,
-    enableDownVotes: Boolean,
-    viewSource: Boolean,
-    postActionBarMode: PostActionBarMode,
-    fromPostActivity: Boolean,
-    scope: CoroutineScope,
-) {
-    val ctx = LocalContext.current
-    var showMoreOptions by remember { mutableStateOf(false) }
-
-    if (showMoreOptions) {
-        val fallbackModerators = remember(moderators) {
-            moderators ?: simulateModerators(
-                ctx = ctx,
-                account = account,
-                forCommunity = postView.community.id,
-            )
-        }
-
-        val amMod = remember(moderators) {
-            amMod(
-                moderators = fallbackModerators,
-                myId = account.id,
-            )
-        }
-
-        val canMod = remember(admins, moderators) {
-            canMod(
-                creatorId = postView.creator.id,
-                admins = admins,
-                moderators = fallbackModerators,
-                myId = account.id,
-            )
-        }
-
-        PostOptionsDropdown(
-            postView = postView,
-            onDismissRequest = { showMoreOptions = false },
-            onCommunityClick = onCommunityClick,
-            onPersonClick = onPersonClick,
-            onEditPostClick = onEditPostClick,
-            onDeletePostClick = onDeletePostClick,
-            onHidePostClick = onHidePostClick,
-            onReportClick = onReportClick,
-            onRemoveClick = onRemoveClick,
-            onBanPersonClick = onBanPersonClick,
-            onBanFromCommunityClick = onBanFromCommunityClick,
-            onLockPostClick = onLockPostClick,
-            onFeaturePostClick = onFeaturePostClick,
-            onViewVotesClick = onViewVotesClick,
-            onViewSourceClick = onViewSourceClick,
-            isCreator = account.id == postView.creator.id,
-            canMod = canMod,
-            amAdmin = account.isAdmin,
-            amMod = amMod,
-            viewSource = viewSource,
-            showViewSource = fromPostActivity,
-            scope = scope,
-        )
-    }
-
-    val horizontalArrangement =
-        when (postActionBarMode) {
-            PostActionBarMode.Long -> Arrangement.spacedBy(XXL_PADDING)
-            PostActionBarMode.LeftHandShort -> Arrangement.spacedBy(LARGE_PADDING)
-            PostActionBarMode.RightHandShort -> Arrangement.spacedBy(LARGE_PADDING)
-        }
-
-    Row(
-        horizontalArrangement = horizontalArrangement,
-        verticalAlignment = Alignment.Bottom,
-        modifier =
-            modifier
-                .fillMaxWidth()
-                .padding(bottom = SMALL_PADDING),
-    ) {
-        // Right handside shows the comments on the left side
-        if (postActionBarMode == PostActionBarMode.RightHandShort) {
-            CommentNewCountRework(
-                comments = postView.counts.comments,
-                unreadCount = postView.unread_comments,
-                account = account,
-                modifier = Modifier.weight(1F, true),
-            )
-        }
-
-        VoteGeneric(
-            myVote = instantScores.myVote,
-            type = VoteType.Upvote,
-            onVoteClick = onUpvoteClick,
-            account = account,
-        )
-        if (enableDownVotes) {
-            VoteGeneric(
-                myVote = instantScores.myVote,
-                type = VoteType.Downvote,
-                onVoteClick = onDownvoteClick,
-                account = account,
-            )
-        }
-
-        if (postActionBarMode == PostActionBarMode.Long) {
-            CommentNewCountRework(
-                comments = postView.counts.comments,
-                unreadCount = postView.unread_comments,
-                account = account,
-                modifier = Modifier.weight(1F, true),
-            )
-        }
-
-        if (showReply) {
-            ActionBarButton(
-                icon = Icons.AutoMirrored.Outlined.Comment,
-                contentDescription = stringResource(R.string.postListing_reply),
-                onClick = { onReplyClick(postView) },
-                account = account,
-            )
-        }
-        SavedButton(
-            saved = postView.saved,
-            account = account,
-            onSaveClick = { onSaveClick(postView) },
-        )
-        ActionBarButton(
-            icon = Icons.Outlined.MoreVert,
-            contentDescription = stringResource(R.string.moreOptions),
-            account = account,
-            onClick = { showMoreOptions = !showMoreOptions },
-            requiresAccount = false,
-            modifier = if (postActionBarMode == PostActionBarMode.LeftHandShort) {
-                Modifier.weight(
-                    1F,
-                    true,
-                )
-            } else {
-                Modifier
-            },
-        )
-
-        if (postActionBarMode == PostActionBarMode.LeftHandShort) {
-            CommentNewCountRework(
-                comments = postView.counts.comments,
-                unreadCount = postView.unread_comments,
-                account = account,
-            )
-        }
-    }
-}
-
-@Composable
-fun SavedButton(
-    saved: Boolean,
-    account: Account,
-    onSaveClick: () -> Unit,
-) {
-    ActionBarButton(
-        icon = if (saved) {
-            Icons.Filled.Bookmark
-        } else {
-            Icons.Outlined.BookmarkBorder
-        },
-        contentDescription = if (saved) {
-            stringResource(R.string.removeBookmark)
-        } else {
-            stringResource(R.string.addBookmark)
-        },
-        onClick = onSaveClick,
-        contentColor = if (saved) {
-            MaterialTheme.colorScheme.primary
-        } else {
-            MaterialTheme.colorScheme.outline
-        },
-        account = account,
-    )
-}
-
-@Composable
-fun CommentNewCountRework(
-    comments: Long,
-    unreadCount: Long,
-    account: Account,
-    modifier: Modifier = Modifier,
-) {
-    val unread =
-        if (unreadCount == 0L || comments == unreadCount) {
-            null
-        } else {
-            (if (unreadCount > 0) "+" else "") + siFormat(unreadCount)
-        }
-
-    ActionBarButtonAndBadge(
-        icon = Icons.Outlined.ChatBubbleOutline,
-        iconBadgeCount = unread,
-        contentDescription = null,
-        text = siFormat(comments),
-        noClick = true,
-        account = account,
-        onClick = {},
-        modifier = modifier,
-    )
-}
-
-@Composable
-fun CommentNewCount(
-    comments: Long,
-    unreadCount: Long,
-    style: TextStyle = MaterialTheme.typography.labelSmall.copy(fontStyle = FontStyle.Italic),
-    spacing: Dp = 0.dp,
-) {
-    val unread =
-        if (unreadCount == 0L || comments == unreadCount) {
-            null
-        } else {
-            unreadCount
-        }
-    if (unread != null) {
-        Spacer(Modifier.padding(horizontal = spacing))
-
-        Text(
-            text = stringResource(R.string.post_listing_new, unread),
-            style = style,
-            color = MaterialTheme.colorScheme.outline,
-        )
-    }
-}
-
-@Preview
-@Composable
-fun CommentCountPreview() {
-    CommentNewCountRework(42, 0, account = AnonAccount)
-}
-
-@Preview
-@Composable
-fun PostFooterLinePreview() {
-    val postView = samplePostView
-    val instantScores =
-        InstantScores(
-            myVote = postView.my_vote,
-            score = postView.counts.score,
-            upvotes = postView.counts.upvotes,
-            downvotes = postView.counts.downvotes,
-        )
-    PostFooterLine(
-        postView = postView,
-        admins = emptyList(),
-        moderators = emptyList(),
-        instantScores = instantScores,
-        onUpvoteClick = {},
-        onDownvoteClick = {},
-        onReplyClick = {},
-        onSaveClick = {},
-        onEditPostClick = {},
-        onDeletePostClick = {},
-        onHidePostClick = {},
-        onReportClick = {},
-        onRemoveClick = {},
-        onBanPersonClick = {},
-        onBanFromCommunityClick = {},
-        onLockPostClick = {},
-        onFeaturePostClick = {},
-        onViewVotesClick = {},
-        onCommunityClick = {},
-        onPersonClick = {},
-        onViewSourceClick = {},
-        account = AnonAccount,
-        enableDownVotes = true,
-        viewSource = false,
-        postActionBarMode = PostActionBarMode.Long,
-        fromPostActivity = true,
-        scope = rememberCoroutineScope(),
-    )
-}
-
-@Preview
-@Composable
-fun PreviewPostListingCard() {
-    PostListing(
-        postView = samplePostView,
-        admins = emptyList(),
-        moderators = emptyList(),
-        useCustomTabs = false,
-        usePrivateTabs = false,
-        onUpvoteClick = {},
-        onDownvoteClick = {},
-        onReplyClick = {},
-        onPostClick = {},
-        onSaveClick = {},
-        onCommunityClick = {},
-        onEditPostClick = {},
-        onDeletePostClick = {},
-        onHidePostClick = {},
-        onReportClick = {},
-        onRemoveClick = {},
-        onBanPersonClick = {},
-        onBanFromCommunityClick = {},
-        onLockPostClick = {},
-        onFeaturePostClick = {},
-        onViewVotesClick = {},
-        onPersonClick = {},
-        fullBody = false,
-        account = AnonAccount,
-        postViewMode = PostViewMode.Card,
-        showVotingArrowsInListView = true,
-        enableDownVotes = true,
-        showAvatar = true,
-        blurNSFW = BlurNSFW.NSFW,
-        appState = rememberJerboaAppState(),
-        showPostLinkPreview = true,
-        showIfRead = true,
-        voteDisplayMode = LocalUserVoteDisplayMode.default(),
-        postActionBarMode = PostActionBarMode.Long,
-        swipeToActionPreset = SwipeToActionPreset.TwoSides,
-    )
-}
-
-@Preview
-@Composable
-fun PreviewLinkPostListing() {
-    PostListing(
-        postView = sampleLinkPostView,
-        admins = emptyList(),
-        moderators = emptyList(),
-        useCustomTabs = false,
-        usePrivateTabs = false,
-        onUpvoteClick = {},
-        onDownvoteClick = {},
-        onReplyClick = {},
-        onPostClick = {},
-        onSaveClick = {},
-        onCommunityClick = {},
-        onEditPostClick = {},
-        onDeletePostClick = {},
-        onHidePostClick = {},
-        onReportClick = {},
-        onRemoveClick = {},
-        onBanPersonClick = {},
-        onBanFromCommunityClick = {},
-        onLockPostClick = {},
-        onFeaturePostClick = {},
-        onViewVotesClick = {},
-        onPersonClick = {},
-        fullBody = false,
-        account = AnonAccount,
-        postViewMode = PostViewMode.Card,
-        showVotingArrowsInListView = true,
-        enableDownVotes = true,
-        showAvatar = true,
-        blurNSFW = BlurNSFW.NSFW,
-        appState = rememberJerboaAppState(),
-        showPostLinkPreview = true,
-        showIfRead = true,
-        voteDisplayMode = LocalUserVoteDisplayMode.default(),
-        postActionBarMode = PostActionBarMode.Long,
-        swipeToActionPreset = SwipeToActionPreset.TwoSides,
-    )
-}
-
-@Preview
-@Composable
-fun PreviewImagePostListingCard() {
-    PostListing(
-        postView = sampleImagePostView,
-        admins = emptyList(),
-        moderators = emptyList(),
-        useCustomTabs = false,
-        usePrivateTabs = false,
-        onUpvoteClick = {},
-        onDownvoteClick = {},
-        onReplyClick = {},
-        onPostClick = {},
-        onSaveClick = {},
-        onCommunityClick = {},
-        onEditPostClick = {},
-        onDeletePostClick = {},
-        onHidePostClick = {},
-        onReportClick = {},
-        onRemoveClick = {},
-        onBanPersonClick = {},
-        onBanFromCommunityClick = {},
-        onLockPostClick = {},
-        onFeaturePostClick = {},
-        onViewVotesClick = {},
-        onPersonClick = {},
-        fullBody = false,
-        account = AnonAccount,
-        postViewMode = PostViewMode.Card,
-        showVotingArrowsInListView = true,
-        enableDownVotes = true,
-        showAvatar = true,
-        blurNSFW = BlurNSFW.NSFW,
-        appState = rememberJerboaAppState(),
-        showPostLinkPreview = true,
-        showIfRead = true,
-        voteDisplayMode = LocalUserVoteDisplayMode.default(),
-        postActionBarMode = PostActionBarMode.Long,
-        swipeToActionPreset = SwipeToActionPreset.TwoSides,
-    )
-}
-
-@Preview
-@Composable
-fun PreviewImagePostListingSmallCard() {
-    PostListing(
-        postView = sampleImagePostView,
-        admins = emptyList(),
-        moderators = emptyList(),
-        useCustomTabs = false,
-        usePrivateTabs = false,
-        onUpvoteClick = {},
-        onDownvoteClick = {},
-        onReplyClick = {},
-        onPostClick = {},
-        onSaveClick = {},
-        onCommunityClick = {},
-        onEditPostClick = {},
-        onDeletePostClick = {},
-        onHidePostClick = {},
-        onReportClick = {},
-        onRemoveClick = {},
-        onBanPersonClick = {},
-        onBanFromCommunityClick = {},
-        onLockPostClick = {},
-        onFeaturePostClick = {},
-        onViewVotesClick = {},
-        onPersonClick = {},
-        fullBody = false,
-        account = AnonAccount,
-        postViewMode = PostViewMode.SmallCard,
-        showVotingArrowsInListView = true,
-        enableDownVotes = true,
-        showAvatar = true,
-        blurNSFW = BlurNSFW.NSFW,
-        appState = rememberJerboaAppState(),
-        showPostLinkPreview = true,
-        showIfRead = true,
-        voteDisplayMode = LocalUserVoteDisplayMode.default(),
-        postActionBarMode = PostActionBarMode.Long,
-        swipeToActionPreset = SwipeToActionPreset.TwoSides,
-    )
-}
-
-@Preview
-@Composable
-fun PreviewLinkNoThumbnailPostListing() {
-    PostListing(
-        postView = sampleLinkNoThumbnailPostView,
-        admins = emptyList(),
-        moderators = emptyList(),
-        useCustomTabs = false,
-        usePrivateTabs = false,
-        onUpvoteClick = {},
-        onDownvoteClick = {},
-        onReplyClick = {},
-        onPostClick = {},
-        onSaveClick = {},
-        onCommunityClick = {},
-        onEditPostClick = {},
-        onDeletePostClick = {},
-        onHidePostClick = {},
-        onReportClick = {},
-        onRemoveClick = {},
-        onBanPersonClick = {},
-        onBanFromCommunityClick = {},
-        onLockPostClick = {},
-        onFeaturePostClick = {},
-        onViewVotesClick = {},
-        onPersonClick = {},
-        fullBody = false,
-        account = AnonAccount,
-        postViewMode = PostViewMode.Card,
-        showVotingArrowsInListView = true,
-        enableDownVotes = true,
-        showAvatar = true,
-        blurNSFW = BlurNSFW.NSFW,
-        appState = rememberJerboaAppState(),
-        showPostLinkPreview = true,
-        showIfRead = true,
-        voteDisplayMode = LocalUserVoteDisplayMode.default(),
-        postActionBarMode = PostActionBarMode.Long,
-        swipeToActionPreset = SwipeToActionPreset.TwoSides,
-    )
-}
-
+@ExperimentalLayoutApi
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun PostListing(
@@ -1295,257 +284,9 @@ fun PostListing(
     }
 }
 
-@Composable
-fun PostVotingTile(
-    instantScores: InstantScores,
-    onUpvoteClick: () -> Unit,
-    onDownvoteClick: () -> Unit,
-    account: Account,
-    enableDownVotes: Boolean,
-    voteDisplayMode: LocalUserVoteDisplayMode,
-) {
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.spacedBy(SMALL_PADDING),
-        modifier =
-            Modifier
-                .fillMaxHeight()
-                .padding(end = MEDIUM_PADDING),
-    ) {
-        VoteGeneric(
-            myVote = instantScores.myVote,
-            type = VoteType.Upvote,
-            onVoteClick = onUpvoteClick,
-            account = account,
-        )
-
-        val scoreOrPctStr = instantScores.scoreOrPctStr(voteDisplayMode)
-
-        Text(
-            text = scoreOrPctStr ?: "",
-            style = MaterialTheme.typography.bodyMedium,
-            color = scoreColor(myVote = instantScores.myVote),
-            // Hide the vote number if its
-            modifier = Modifier.alpha(if (scoreOrPctStr != null) 1f else 0f),
-        )
-
-        if (enableDownVotes) {
-            // invisible Text below aligns width of PostVotingTiles
-            Text(
-                text = "00000",
-                modifier = Modifier.height(0.dp),
-                style = MaterialTheme.typography.bodyMedium,
-            )
-            VoteGeneric(
-                myVote = instantScores.myVote,
-                type = VoteType.Downvote,
-                onVoteClick = onDownvoteClick,
-                account = account,
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-fun PostListingList(
-    postView: PostView,
-    instantScores: InstantScores,
-    onUpvoteClick: () -> Unit,
-    onDownvoteClick: () -> Unit,
-    onPostClick: (postView: PostView) -> Unit,
-    showCommunityName: Boolean = true,
-    account: Account,
-    showVotingArrowsInListView: Boolean,
-    useCustomTabs: Boolean,
-    usePrivateTabs: Boolean,
-    blurNSFW: BlurNSFW,
-    appState: JerboaAppState,
-    showIfRead: Boolean,
-    enableDownVotes: Boolean,
-    voteDisplayMode: LocalUserVoteDisplayMode,
-) {
-    Column(
-        modifier =
-            Modifier
-                .padding(
-                    horizontal = MEDIUM_PADDING,
-                    vertical = MEDIUM_PADDING,
-                ).testTag("jerboa:post"),
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement =
-                Arrangement.spacedBy(
-                    SMALL_PADDING,
-                ),
-        ) {
-            if (showVotingArrowsInListView) {
-                PostVotingTile(
-                    instantScores = instantScores,
-                    onUpvoteClick = onUpvoteClick,
-                    onDownvoteClick = onDownvoteClick,
-                    account = account,
-                    enableDownVotes = enableDownVotes,
-                    voteDisplayMode = voteDisplayMode,
-                )
-            }
-            Column(
-                modifier =
-                    Modifier
-                        .weight(1f)
-                        .clickable { onPostClick(postView) },
-                verticalArrangement = Arrangement.spacedBy(SMALL_PADDING),
-            ) {
-                PostName(post = postView.post, read = postView.read, showIfRead = showIfRead)
-                FlowRow(
-                    horizontalArrangement = Arrangement.spacedBy(SMALLER_PADDING, Alignment.Start),
-                ) {
-                    // You must use a center align modifier for each of these
-                    val centerMod = Modifier.align(Alignment.CenterVertically)
-                    if (showCommunityName) {
-                        CommunityLink(
-                            community = postView.community,
-                            onClick = {},
-                            clickable = false,
-                            showDefaultIcon = false,
-                            showAvatar = false,
-                            blurNSFW = blurNSFW,
-                            modifier = centerMod,
-                        )
-                        DotSpacer(modifier = centerMod)
-                    }
-                    PersonProfileLink(
-                        person = postView.creator,
-                        onClick = {},
-                        clickable = false,
-                        color = MaterialTheme.colorScheme.outline,
-                        showAvatar = false,
-                        modifier = centerMod,
-                    )
-                    DotSpacer(modifier = centerMod)
-                    postView.post.url?.also { postUrl ->
-                        if (!isSameInstance(postUrl, account.instance)) {
-                            val hostName = hostName(postUrl)
-                            hostName?.also {
-                                Text(
-                                    text = it,
-                                    color = MaterialTheme.colorScheme.outline,
-                                    style = MaterialTheme.typography.labelMedium,
-                                    fontFamily = FontFamily.Monospace,
-                                    modifier = centerMod,
-                                )
-                                DotSpacer(modifier = centerMod)
-                            }
-                        }
-                    }
-                    TimeAgo(
-                        published = postView.post.published,
-                        updated = postView.post.updated,
-                        modifier = centerMod,
-                    )
-                }
-                Row(
-                    horizontalArrangement = Arrangement.spacedBy(SMALLER_PADDING),
-                    verticalAlignment = Alignment.CenterVertically,
-                ) {
-                    if (!showVotingArrowsInListView) {
-                        Text(
-                            text = instantScores.score.toString(),
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = scoreColor(myVote = instantScores.myVote),
-                        )
-                        DotSpacer()
-                    }
-                    Text(
-                        text =
-                            stringResource(
-                                R.string.post_listing_comments_count,
-                                postView.counts.comments,
-                            ),
-                        style = MaterialTheme.typography.labelMedium,
-                        color = MaterialTheme.colorScheme.outline,
-                    )
-                    CommentNewCount(
-                        comments = postView.counts.comments,
-                        unreadCount = postView.unread_comments,
-                        style = MaterialTheme.typography.labelMedium,
-                    )
-                    NsfwBadge(nsfwCheck(postView))
-                }
-            }
-            ThumbnailTile(
-                post = postView.post,
-                useCustomTabs = useCustomTabs,
-                usePrivateTabs = usePrivateTabs,
-                blurEnabled = blurNSFW.needBlur(postView),
-                appState = appState,
-            )
-        }
-    }
-}
-
-@Preview
-@Composable
-fun PostListingListPreview() {
-    val postView = samplePostView
-    val instantScores =
-        InstantScores(
-            myVote = postView.my_vote,
-            score = postView.counts.score,
-            upvotes = postView.counts.upvotes,
-            downvotes = postView.counts.downvotes,
-        )
-    PostListingList(
-        postView = postView,
-        instantScores = instantScores,
-        onUpvoteClick = {},
-        onDownvoteClick = {},
-        onPostClick = {},
-        account = AnonAccount,
-        showVotingArrowsInListView = true,
-        useCustomTabs = false,
-        usePrivateTabs = false,
-        blurNSFW = BlurNSFW.NSFW,
-        appState = rememberJerboaAppState(),
-        showIfRead = true,
-        enableDownVotes = false,
-        voteDisplayMode = LocalUserVoteDisplayMode.default(),
-    )
-}
-
-@Preview
-@Composable
-fun PostListingListWithThumbPreview() {
-    val postView = sampleImagePostView
-    val instantScores =
-        InstantScores(
-            myVote = postView.my_vote,
-            score = postView.counts.score,
-            upvotes = postView.counts.upvotes,
-            downvotes = postView.counts.downvotes,
-        )
-    PostListingList(
-        postView = postView,
-        instantScores = instantScores,
-        onUpvoteClick = {},
-        onDownvoteClick = {},
-        onPostClick = {},
-        account = AnonAccount,
-        showVotingArrowsInListView = true,
-        useCustomTabs = false,
-        usePrivateTabs = false,
-        blurNSFW = BlurNSFW.NSFW,
-        appState = rememberJerboaAppState(),
-        showIfRead = true,
-        enableDownVotes = false,
-        voteDisplayMode = LocalUserVoteDisplayMode.default(),
-    )
-}
-
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
-private fun ThumbnailTile(
+fun ThumbnailTile(
     post: Post,
     useCustomTabs: Boolean,
     usePrivateTabs: Boolean,
@@ -1622,136 +363,6 @@ private fun ThumbnailTile(
 }
 
 @Composable
-fun PostListingCard(
-    postView: PostView,
-    admins: List<PersonView>,
-    moderators: List<PersonId>?,
-    instantScores: InstantScores,
-    onUpvoteClick: () -> Unit,
-    onDownvoteClick: () -> Unit,
-    onReplyClick: (postView: PostView) -> Unit = {},
-    onPostClick: (postView: PostView) -> Unit,
-    onSaveClick: (postView: PostView) -> Unit,
-    onCommunityClick: (community: Community) -> Unit,
-    onEditPostClick: (postView: PostView) -> Unit,
-    onDeletePostClick: (postView: PostView) -> Unit,
-    onHidePostClick: (postView: PostView) -> Unit,
-    onReportClick: (postView: PostView) -> Unit,
-    onRemoveClick: (postView: PostView) -> Unit,
-    onBanPersonClick: (person: Person) -> Unit,
-    onBanFromCommunityClick: (banData: BanFromCommunityData) -> Unit,
-    onLockPostClick: (postView: PostView) -> Unit,
-    onFeaturePostClick: (data: PostFeatureData) -> Unit,
-    onPersonClick: (personId: PersonId) -> Unit,
-    onViewVotesClick: (PostId) -> Unit,
-    onViewSourceClick: () -> Unit,
-    viewSource: Boolean,
-    showReply: Boolean = false,
-    showCommunityName: Boolean = true,
-    fullBody: Boolean,
-    account: Account,
-    expandedImage: Boolean,
-    enableDownVotes: Boolean,
-    showAvatar: Boolean,
-    useCustomTabs: Boolean,
-    usePrivateTabs: Boolean,
-    blurNSFW: BlurNSFW,
-    showPostLinkPreview: Boolean,
-    appState: JerboaAppState,
-    showIfRead: Boolean = false,
-    voteDisplayMode: LocalUserVoteDisplayMode,
-    postActionBarMode: PostActionBarMode,
-) {
-    Column(
-        modifier =
-            Modifier
-                .padding(vertical = MEDIUM_PADDING)
-                .clickable { onPostClick(postView) }
-                .testTag("jerboa:post"),
-        // see https://stackoverflow.com/questions/77010371/prevent-popup-from-adding-padding-in-a-column-with-arrangement-spacedbylarge-p
-        // verticalArrangement = Arrangement.spacedBy(MEDIUM_PADDING),
-    ) {
-        // Header
-        PostHeaderLine(
-            post = postView.post,
-            creator = postView.creator,
-            community = postView.community,
-            creatorBannedFromCommunity = postView.creator_banned_from_community,
-            instantScores = instantScores,
-            onCommunityClick = onCommunityClick,
-            onPersonClick = onPersonClick,
-            showCommunityName = showCommunityName,
-            modifier = Modifier.padding(horizontal = MEDIUM_PADDING),
-            showAvatar = showAvatar,
-            blurNSFW = blurNSFW,
-            voteDisplayMode = voteDisplayMode,
-            fullBody = fullBody,
-        )
-
-        Spacer(modifier = Modifier.padding(vertical = MEDIUM_PADDING))
-
-        //  Title + metadata
-        PostBody(
-            post = postView.post,
-            read = postView.read,
-            fullBody = fullBody,
-            viewSource = viewSource,
-            expandedImage = expandedImage,
-            account = account,
-            useCustomTabs = useCustomTabs,
-            usePrivateTabs = usePrivateTabs,
-            blurEnabled = blurNSFW.needBlur(postView),
-            showPostLinkPreview = showPostLinkPreview,
-            appState = appState,
-            clickBody = { onPostClick(postView) },
-            showIfRead = showIfRead,
-        )
-
-        Spacer(modifier = Modifier.padding(vertical = MEDIUM_PADDING))
-
-        // Footer bar
-        PostFooterLine(
-            postView = postView,
-            admins = admins,
-            moderators = moderators,
-            instantScores = instantScores,
-            onUpvoteClick = onUpvoteClick,
-            onDownvoteClick = onDownvoteClick,
-            onReplyClick = onReplyClick,
-            onSaveClick = onSaveClick,
-            onEditPostClick = onEditPostClick,
-            onDeletePostClick = onDeletePostClick,
-            onHidePostClick = onHidePostClick,
-            onReportClick = onReportClick,
-            onRemoveClick = onRemoveClick,
-            onBanPersonClick = onBanPersonClick,
-            onBanFromCommunityClick = onBanFromCommunityClick,
-            onLockPostClick = onLockPostClick,
-            onFeaturePostClick = onFeaturePostClick,
-            onViewVotesClick = onViewVotesClick,
-            onCommunityClick = onCommunityClick,
-            onPersonClick = onPersonClick,
-            onViewSourceClick = onViewSourceClick,
-            modifier = Modifier.padding(horizontal = MEDIUM_PADDING),
-            showReply = showReply,
-            account = account,
-            enableDownVotes = enableDownVotes,
-            viewSource = viewSource,
-            postActionBarMode = postActionBarMode,
-            fromPostActivity = fullBody,
-            scope = appState.coroutineScope,
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Preview
-@Composable
-fun PostListingHeaderPreview() {
-    SimpleTopAppBar("Post", onClickBack = {})
-}
-
-@Composable
 fun MetadataCard(post: Post) {
     val embedTitle = post.embed_title
     // If there is a valid title or description show it
@@ -1806,5 +417,30 @@ fun TorrentHelpInfo() {
                 )
             }
         },
+    )
+}
+
+@Composable
+fun PostName(
+    post: Post,
+    read: Boolean,
+    showIfRead: Boolean,
+) {
+    val color =
+        if (showIfRead && read) {
+            MaterialTheme.colorScheme.outline
+        } else if (post.featured_local) {
+            MaterialTheme.colorScheme.primary
+        } else if (post.featured_community) {
+            MaterialTheme.colorScheme.secondary
+        } else {
+            Color.Unspecified
+        }
+
+    Text(
+        text = post.name,
+        style = MaterialTheme.typography.headlineMedium,
+        color = color,
+        modifier = Modifier.testTag("jerboa:posttitle"),
     )
 }
