+++ app/src/main/java/com/jerboa/ui/components/post/PostScreen.kt
@@ -4,6 +4,7 @@ import android.util.Log
 import androidx.compose.foundation.focusable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -109,6 +110,7 @@ object PostViewReturn {
     const val POST_VIEW = "post-view::return(post-view)"
 }
 
+@ExperimentalLayoutApi
 @OptIn(
     ExperimentalMaterial3Api::class,
     ExperimentalComposeUiApi::class,
