+++ app/src/main/java/com/jerboa/ui/theme/Sizes.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.unit.dp
 
 const val DEFAULT_FONT_SIZE = 16
 
-val ACTION_BAR_ICON_SIZE = 16.dp
+val ACTION_BAR_ICON_SIZE = 12.dp
 val MARKDOWN_BAR_ICON_SIZE = 24.dp
 
 val BORDER_WIDTH = 1.dp
@@ -16,7 +16,6 @@ val XL_PADDING = 16.dp
 val XXL_PADDING = 20.dp
 
 val ICON_SIZE = 20.dp
-val MEDIUM_ICON_SIZE = 48.dp
 val LARGER_ICON_SIZE = 80.dp
 val DRAWER_BANNER_SIZE = 96.dp
 val PROFILE_BANNER_SIZE = 128.dp
